Better extensibility of the effect system
The on-collision damage system introduced in the previous commit (1cd0ba9) is not
extensible enough.

For example, adding a freezing projectile would have required to add a
IFreezingEntity class, containing an inflictFreeze method that all freezing
entities must implement. This would introduce a lot of duplicate code, especially
if even more effects were added.

Now, the effect inflicting part is entirely managed by the specialized listeners,
they only have to ask the permission for applying their effect to the event source
via the function canApplyEffect defined in the IEffectEntity interface.