refactor: global file chooser and new map file loading
Now that we have an editor, maps come in two types: internal maps (distributed
with the game and read with getResourceAsStream) and external maps (created by
users and read with File objects). Those two types of maps can both be
encapsulated in the Resource class and passed to the model that will process
them exactly the same.