fix: fix obstaclesBetween function
the enemy sometimes stopped before being able to properly shoot the
player, ie. the bullet would hit a wall before the player.
This is because the obstaclesBetween function was only checking
for obstacles in the direct path of the bullet, which didn't account
for edges of obstacles. I changed the function so that it checks
all the surrounding tiles of the path as well (excluding the start/end).